feat: Replicate face registration in teacher's detailed edit view
This commit introduces the functionality for teachers to register a student's face from the "Edit Detailed" view in the student database.

Key changes:
- A reusable `FaceRegistration.jsx` component has been created by extracting the logic from `StudentProfilePage.jsx`.
- `StudentProfilePage.jsx` has been refactored to use the new `FaceRegistration` component.
- `TeacherDatabasePage.jsx` now includes a "Register Face" button in the "Edit Detailed" modal, which opens the `FaceRegistration` component.
- A new backend endpoint `POST /api/admin/students/:studentId/register-face` has been added to allow teachers to save a student's face descriptors.
